refactor(ImageViewer): rename ImageViewerModel to ImageViewerModal
